refactor(test_dispatch_register): split tests for size compliance (W2D)
Resolves OI-1179 (file was 580L, now 399L) and OI-1180 (test_read_events_takes_shared_lock was 44L, now 16L via extracted helper).

- Moved path-resolution and shared-lock tests to tests/test_dispatch_register_path.py (220L)
- Extracted _run_exclusive_write_then_read helper to reduce lock test to 16L
- All 35 tests preserved; no source-code changes

Dispatch-ID: 20260501-w2d-test-dispatch-register
